backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
aiofiles
python-dotenv
langgraph
langchain
langchain-ollama
langchain-community
chromadb
rank-bm25
duckduckgo-search
PyPDF2
python-docx
pandas
numpy
python-jose[cryptography]
passlib[bcrypt]

pypdf